requirements.txt
google-cloud-storage
langchain
langchain-openai
python-dotenv
langchain-community
pypdf
chromadb
flask
